Compressed pages together sql and php work
since all pages were pretty similar i compresssed them together and began work onit as a 1 page site instead. added a bunch to the pia site sql. added accessing php requests and functions for the sql data added and did the display for the data added in the index page.